Move get_table() into a common lib module
This will also be used by the copy script to create table instances.
