+++ home/bin/aws_dynamodb_truncate.py
@@ -2,6 +2,8 @@
 
 from typing import List, Optional, cast
 
+from lib.aws.dynamodb import get_table
+
 
 def main() -> int:
     args = get_parser().parse_args()
@@ -85,22 +87,6 @@ def get_parser():
     )
 
     return parser
-
-
-def get_table(
-    profile: str,
-    region: str,
-    table_name: str,
-    retries: Optional[int],
-):
-    from boto3 import Session
-    from botocore.config import Config
-
-    session = Session(profile_name=profile, region_name=region)
-    config = Config(retries={"max_attempts": retries}) if retries else Config()
-    dynamodb = session.resource("dynamodb", config=config)
-
-    return dynamodb.Table(table_name)
 
 
 def get_item_pages(table, consistent_scan: bool, scan_size: Optional[int]):
